refact(order): 포인트 BigDecimal -> Long

+++ order-service/src/main/java/com/rushcrew/order_service/application/query/dto/OrderDetailDto.java
@@ -24,7 +24,7 @@ public class OrderDetailDto {
 	private Long userId;
 	private String orderStatus;
 	private BigDecimal totalAmount;
-	private BigDecimal pointUsed;
+	private Long pointUsed;
 	private BigDecimal finalAmount;
 	private Instant orderedAt;
 	private Instant paymentCompletedAt;
